+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add `Rapa::Resources::ItemResource#ebook?`
+
 ## 0.4.2
 
 - Add `:browse_node_id` options to `Rapa::Client#search_items`